Improved how numbers are shown
Refactored metric formatting into a reusable utility function (`client/src/lib/formatters.ts`) and removed redundant code from `client/src/components/labs/lab-results-analyzer.tsx` and `client/src/components/labs/metric-analysis-card.tsx`.  Also removed authentication middleware from `/analyze-metric` and `/analyze-recommendation` endpoints in `server/routes/metricsAnalysisRoutes.ts`.
